Fix ESLint errors in frontend code
- Add PropTypes validation to Gallery, ProtectedRoute, and AuthProvider components
- Install prop-types package as dev dependency
- Fix vite.config.js Node.js globals issue by updating ESLint config
- Separate AuthContext from useAuth hook to resolve Fast Refresh warning
- Update all import paths to use new useAuth hook location
- Fix test mock imports to match new hook structure

Resolves all 9 ESLint errors and 1 warning in CI pipeline.
