+++ frontend/package.json
@@ -40,6 +40,7 @@
     "jsdom": "^26.0.0",
     "lucide-react": "^0.477.0",
     "postcss": "^8.5.3",
+    "prop-types": "^15.8.1",
     "tailwindcss": "^3.4.17",
     "vite": "^6.0.5",
     "vitest": "^3.0.7"
