+++ frontend/src/components/Navbar.jsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../hooks/useAuth";
 
 const Navbar = () => {
   const { isAuthenticated, userType, logout } = useAuth();
