+++ frontend/src/context/AuthContextProvider.jsx
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const AuthContext = createContext(null);
