+++ frontend/src/pages/Booking/Booking.jsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import bookingApi from "../../api/bookingApi";
-import { useAuth } from "../../context/AuthContext";
+import { useAuth } from "../../hooks/useAuth";
 
 function Booking() {
   const { photographer_id } = useParams();
