+++ frontend/src/pages/BookingManager.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import bookingApi from "../api/bookingApi";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../hooks/useAuth";
 
 export default function BookingManager() {
   const { photographerId } = useAuth();
