+++ frontend/src/pages/Home/Home.jsx
@@ -1,4 +1,4 @@
-import { useAuth } from "../../context/AuthContext";
+import { useAuth } from "../../hooks/useAuth";
 import { Link } from "react-router-dom";
 
 export default function Home() {
