+++ frontend/src/pages/Login/Login.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useNavigate, Link } from "react-router-dom";
-import { useAuth } from "../../context/AuthContext";
+import { useAuth } from "../../hooks/useAuth";
 import { loginUser } from "../../api/client";
 
 export default function Login() {
