+++ frontend/src/pages/PhotographerDashboard.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState, useRef } from "react";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../hooks/useAuth";
 import { deletePhoto, uploadPhoto } from "../api/client";
 import photoApi from "../api/photoApi";
 import { PlusCircle, Trash2 } from "lucide-react";
