+++ frontend/vite.config.js
@@ -1,10 +1,10 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
+import dotenv from "dotenv";
 
 const mode = process.env.NODE_ENV || "development";
 const envFile = `.env.${mode}`;
 
-import dotenv from "dotenv";
 dotenv.config({ path: envFile });
 
 // https://vite.dev/config/
@@ -21,7 +21,7 @@ export default defineConfig({
   },
   define: {
     "import.meta.env.VITE_API_BASE_URL": JSON.stringify(
-      process.env.VITE_API_BASE_URL
+      process.env.VITE_API_BASE_URL || "http://localhost:5002"
     ),
   },
 });
